Create sms-in.tcl
This version does not support multiple threads as it doesn't handle that an SMS changed index in the local SMS register. If we are to support >1 thread, we need to ensure the correct SMS is deleted.
Until its clear how the NotificationGateway App device will function, we are are hardcoding feedback value to 200.